+++ browser-features/chrome/common/panel-sidebar/components/panel-sidebar-modal.tsx
@@ -197,7 +197,7 @@ export class PanelSidebarAddModal {
         label: texts.url,
         value: globalThis.gBrowser.currentURI?.spec ?? "",
         required: true,
-        placeholder: "https://floorp.app",
+        placeholder: "https://stratus-browser.org",
         when: { id: "type", value: "web" },
         onInput: (value: string) => {
           if (value) {